+++ internal/model/game_config.go
@@ -26,10 +26,10 @@ type GameConfig struct {
 	Status int8 `gorm:"column:status" json:"status"`
 
 	// 创建时间
-	CreateTime uint32 `gorm:"column:create_time" json:"create_time"`
+	CreateTime int64 `gorm:"column:create_time" json:"create_time"`
 
 	// 修改时间
-	UpdateTime uint32 `gorm:"column:update_time" json:"update_time"`
+	UpdateTime int64 `gorm:"column:update_time" json:"update_time"`
 }
 
 // TableName 表名
