+++ internal/model/jackpot_log.go
@@ -29,7 +29,7 @@ type JackpotLog struct {
 	Amount uint64 `gorm:"column:amount" json:"amount"`
 
 	// 中奖时间
-	CreateTime uint32 `gorm:"column:create_time" json:"create_time"`
+	CreateTime int64 `gorm:"column:create_time" json:"create_time"`
 }
 
 // TableName 表名
