+++ internal/model/rollback_log.go
@@ -52,7 +52,7 @@ type RollbackLog struct {
 	Status int8 `gorm:"column:status" json:"status"`
 
 	// 创建时间
-	CreateTime uint32 `gorm:"column:create_time" json:"create_time"`
+	CreateTime int64 `gorm:"column:create_time" json:"create_time"`
 }
 
 // TableName 表名
