+++ internal/model/rtp_stat.go
@@ -18,10 +18,10 @@ type RTPStat struct {
 	TotalWin uint64 `gorm:"column:total_win" json:"total_win"`
 
 	// 创建时间
-	CreateTime uint32 `gorm:"column:create_time" json:"create_time"`
+	CreateTime int64 `gorm:"column:create_time" json:"create_time"`
 
 	// 修改时间
-	UpdateTime uint32 `gorm:"column:update_time" json:"update_time"`
+	UpdateTime int64 `gorm:"column:update_time" json:"update_time"`
 }
 
 // TableName 表名
